feat: implement CLI configuration and runtime management for sensor data processing
